+++ test/builders/assembly_builders_test.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+require_relative '../test_helper'
+require 'core/builder'
+require 'builders/floor/builder'
+require 'builders/wall/builder'
+require 'builders/roof/builder'
+
+class AssemblyBuildersTest < Minitest::Test
+  def setup
+    @container = Object.new
+  end
+
+  def test_primary_builders_have_stable_assembly_identities
+    builders = [
+      ForgeBuild::Builders::Floor::Builder.new(container: @container),
+      ForgeBuild::Builders::Wall::Builder.new(container: @container),
+      ForgeBuild::Builders::Roof::Builder.new(container: @container)
+    ]
+
+    assert_equal %i[floor wall roof], builders.map(&:id)
+    assert_equal ['Floor Builder', 'Wall Builder', 'Roof Builder'], builders.map(&:name)
+  end
+
+  def test_existing_tools_moved_to_correct_assemblies
+    floor = ForgeBuild::Builders::Floor::Builder.new(container: @container)
+    wall = ForgeBuild::Builders::Wall::Builder.new(container: @container)
+
+    assert_equal %i[slab_on_grade equipment_pad], floor.tools.map(&:id)
+    assert_includes wall.tools.map(&:id), :cmu_wall
+    assert_includes wall.tools.map(&:id), :brick_veneer
+  end
+end